+++ lib/routes/index.js
@@ -10,7 +10,7 @@ const errorHandler = require('../util/error-handler')
 const {getCommune} = require('../util/cog')
 const sync = require('../sync')
 const {createHabilitation, sendPinCode, validatePinCode, fetchHabilitation} = require('../habilitations')
-const {checkCadastre} = require('../check-cadastre')
+const {checkCadastre} = require('../cadastre')
 const stats = require('./stats')
 const adminRoutes = require('./admin')
