+++ lib/util/__tests__/check-cadastre.js
@@ -1,5 +1,5 @@
 const test = require('ava')
-const {checkCadastre} = require('../../check-cadastre')
+const {checkCadastre} = require('../../cadastre')
 
 test('Commune with cadastre', t => {
   const communeWithCadastre = checkCadastre('38095')
